Added  supported for hosted background service too.

+++ EduArchivingCullingEnginePOC/Business/Framework/CleanupHandler.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Abstractions.Internal.Framework;
 
 namespace Business.Framework
 {
@@ -7,16 +8,29 @@ namespace Business.Framework
     {
         private readonly InputOutputFilesManager _filesManager;
 
+
         public CleanupHandler(InputOutputFilesManager filesManager)
         {
             _filesManager = filesManager;
+            ArchiveFileCreationHandler.ArchiveFilesCreated += ArchiveFileCreationHandler_ArchiveFilesCreated;
+        }
+
+        private void ArchiveFileCreationHandler_ArchiveFilesCreated(object sender, Abstractions.Internal.Framework.ArchivingEventArgs eventArgs)
+        {
+            foreach (var entity in eventArgs.EntityDataFilesMapper.Keys)
+            {
+                CleanupTemporaryData(eventArgs.EntityDataFilesMapper[entity]);
+            }
         }
 
         public void CleanupTemporaryData(List<string> dataFilesToDelete)
         {
-            foreach (var file in dataFilesToDelete)
+            if (dataFilesToDelete != null && dataFilesToDelete.Count > 0)
             {
-                _filesManager.DeleteInputDataFile(file);
+                foreach (var file in dataFilesToDelete)
+                {
+                    _filesManager.DeleteInputDataFile(file);
+                }
             }
         }
     }
